requirements.txt
# API and Authentication
smartapi-python==1.3.5
pyotp==2.9.0
python-dotenv==1.0.1

# Database
duckdb==0.9.2

# Utilities
logzero==1.7.0
pytz==2024.1
pandas==2.2.0  # Required by smart-api-python for historical data

# Type Hints
typing-extensions==4.9.0 